Fixing RestartAndRemakeController
