+++ Assets/Scripts/UI/RestartRemakeContollers/ExplorationRestartRemakeContoller.cs
@@ -1,0 +1,9 @@
+using Maes;
+
+namespace MAES.UI.RestartRemakeContollers
+{
+    public class ExplorationRestartRemakeContoller : RestartRemakeContoller<ExplorationSimulation>
+    {
+        
+    }
+}
