+++ Assets/Scripts/UI/RestartRemakeContollers/PatrollingRestartRemakeContoller.cs
@@ -1,0 +1,9 @@
+using Maes;
+
+namespace MAES.UI.RestartRemakeContollers
+{
+    public class PatrollingRestartRemakeContoller : RestartRemakeContoller<PatrollingSimulation>
+    {
+        
+    }
+}
