+++ Assets/Scripts/UI/RestartRemakeContollers/RestartRemakeContoller.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 457fc588aa1c1304b9345cb8a62b6b5e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
